# openapi.yaml
openapi: 3.0.3
info:
  title: Transparency Dashboard API
  description: API for tracking grants and spending on-chain for organizational transparency
  version: 1.0.0
  contact:
    name: Ethereum Community Foundation
servers:
  - url: /api/v1
    description: API server
paths:
...
  /api/v1/admins:
    get:
      summary: List admins
      description: Returns a list of all administrators
      operationId: getAdmins
      tags:
        - Admins
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully retrieved admins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Admin'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Add admin
      description: Adds a new administrator
      operationId: addAdmin
      tags:
        - Admins
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Display name of the admin
                  example: "John Doe"
                address:
                  type: string
                  pattern: '^(0x)?[0-9a-f]{40}$'
                  description: Ethereum address of the admin
                  example: "0x742d35cc6670c02bb4bbecbb5fb95e8b7d7b36bf"
              required:
                - name
                - address
      responses:
        '201':
          description: Admin added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Admin'
        '400':
          description: Bad request - invalid input or admin already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admins/{address}:
    delete:
      summary: Remove admin
      description: Removes an administrator by their Ethereum address
      operationId: removeAdmin
      tags:
        - Admins
      security:
        - bearerAuth: []
      parameters:
        - name: address
          in: path
          required: true
          description: The Ethereum address of the admin to remove
          schema:
            type: string
            pattern: '^(0x)?[0-9a-f]{40}$'
            example: "0x742d35cc6670c02bb4bbecbb5fb95e8b7d7b36bf"
      responses:
        '204':
          description: Admin removed successfully
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Admin not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/admin-actions:
    get:
      summary: Get admin actions
      description: Returns a list of actions taken by administrators (audit log)
      operationId: getAdminActions
      tags:
        - Admins
      security:
        - bearerAuth: []
      parameters:
        - name: adminAddress
          in: query
          description: Filter actions by admin address
          required: false
          schema:
            type: string
            pattern: '^(0x)?[0-9a-f]{40}$'
            example: "0x742d35cc6670c02bb4bbecbb5fb95e8b7d7b36bf"
        - name: action
          in: query
          description: Filter by action type
          required: false
          schema:
            type: string
            enum: [create_expense, update_expense, delete_expense, upload_receipt, delete_receipt, create_grant, update_grant, update_milestones, add_admin, remove_admin, update_transfer_party]
        - name: limit
          in: query
          description: Maximum number of actions to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: offset
          in: query
          description: Number of actions to skip for pagination
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Successfully retrieved admin actions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminAction'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Failed to fetch treasury data"
        code:
          type: string
          description: Error code
          example: "TREASURY_FETCH_ERROR"
      required:
        - error
        - code
...
    Admin:
      type: object
      properties:
        name:
          type: string
          description: Display name of the admin
          example: "John Doe"
        address:
          type: string
          pattern: '^(0x)?[0-9a-f]{40}$'
          description: Ethereum address of the admin (serves as unique identifier)
          example: "0x742d35cc6670c02bb4bbecbb5fb95e8b7d7b36bf"
      required:
        - name
        - address
    AdminAction:
      type: object
      properties:
        id:
          type: string
          description: Unique action identifier
          example: "action_12345"
        adminAddress:
          type: string
          pattern: '^(0x)?[0-9a-f]{40}$'
          description: Ethereum address of the admin who performed the action
          example: "0x742d35cc6670c02bb4bbecbb5fb95e8b7d7b36bf"
        adminName:
          type: string
          description: Name of the admin who performed the action
          example: "John Doe"
        action:
          type: string
          enum: [create_expense, update_expense, delete_expense, upload_receipt, delete_receipt, create_grant, update_grant, update_milestones, add_admin, remove_admin, update_transfer_party]
          description: Type of action performed
          example: "create_expense"
        resourceType:
          type: string
          enum: [expense, receipt, grant, admin, transfer_party]
          description: Type of resource the action was performed on
          example: "expense"
        resourceId:
          type: string
          description: ID of the resource that was acted upon
          example: "exp_12345"
        details:
          type: object
          description: Additional details about the action (varies by action type)
          example: {"expenseName": "Flight to conference", "amount": "450.00"}
        timestamp:
          type: string
          format: date-time
          description: When the action was performed
          example: "2023-12-15T10:30:00Z"
      required:
        - id
        - adminAddress
        - adminName
        - action
        - resourceType
        - resourceId
        - timestamp
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...